requirements.txt
numpy
torch
tqdm
matplotlib
python-chess
PyYAML
flask